+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.0.0
 
 - update @vcmap/core and @vcmap/ui to 6.x
+- update state to be not included in create-link when multiView is not active
 
 # v1.1.0