+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - update @vcmap/core and @vcmap/ui to 6.x
 - update state to be not included in create-link when multiView is not active
+- fix bug where views could not be moved by dragging when CesiumMap was selected
 
 # v1.1.0