+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # v2.0.0
 
-- update @vcmap/core and @vcmap/ui to 6.x
-- update state to be not included in create-link when multiView is not active
-- fix bug where views could not be moved by dragging when CesiumMap was selected
+- updates @vcmap/core and @vcmap/ui to 6.x
+- updates state to be not included in create-link when multiView is not active
+- fixes a bug where views could not be moved by dragging when CesiumMap was selected
 
 # v1.1.0